fixed many bugs!

+++ parser/parser.py
@@ -1,4 +1,5 @@
 import re
+import pprint
 
 import requests
 
@@ -7,19 +8,20 @@ import department
 import prereq
 
 
-DEPARTMENT = "([A-Za-z]{3,4})"
+DEPARTMENT = "([A-Za-z]{3,7})"
 
 ABBREV_DICT = {'Computer Science Engineering': 'CSE',
                'Mathematics': 'MATH', }
 
 # Special Course definitions
 SECOND_WRIT_STRS = ['Gen Ed Writing Level 2',
-                    'second level writing course']
+                    'second level writing course',
+                    'Second writing course']
 SECOND_WRIT = {'name': 'Gen Ed Writing Level 2', 'number': 2367}
 
 NUMBER = "(\d{4}(?:\.\d{2})?)"
 # NUMBER = "(\d{4})"
-COURSE = r"(?:(?:{0} )?{1})|{2}|{3}".format(DEPARTMENT, NUMBER, *SECOND_WRIT_STRS)
+COURSE = r"(?:(?:{0} )?{1})|{2}|{3}|{4}".format(DEPARTMENT, NUMBER, *SECOND_WRIT_STRS)
 
 OR_GROUP = 'Or Group'
 
@@ -36,14 +38,14 @@ def search_dept(c):
         return c
 
 
-def add_prereqs(self, prereqs):
-    print("parent: " + self.name)
+def add_prereqs(parent, prereqs):
+    print("parent: " + parent.name)
 
-    if self.name == 'Or Group':
-        g_parent = self.parent.all()[0]
+    if parent.name == 'Or Group':
+        g_parent = parent.parent.all()[0]
         parent_dept = g_parent.dept.all()[0]
     else:
-        parent_dept = self.dept.all()[0]
+        parent_dept = parent.dept.all()[0]
 
     # get the university using the parent's department's uni
     uni = parent_dept.uni.all()[0]
@@ -59,41 +61,40 @@ def add_prereqs(self, prereqs):
         if m_obj.group(1):
             dept_str = m_obj.group(1).upper()
 
-        if dept_str != 'NOT':
-            if m_obj.group(2):
-                number = int(m_obj.group(2))
-                print("dept: " + dept_str + " number: " + str(number))
+        if m_obj.group(2):
+            number = int(m_obj.group(2))
+            print("dept: " + dept_str + " number: " + str(number))
 
-                # check if department already exists in this uni
-                dept_list = uni.dept.search(name=dept_str)
+            # check if department already exists in this uni
+            dept_list = uni.dept.search(name=dept_str)
 
-                if len(dept_list) > 0:
-                    print("department already exists")
-                    # check to see if the course is already listed in the
-                    # department
-                    dept = dept_list[0]
-                    c_list = dept.course.search(number=number)
+            if len(dept_list) > 0:
+                print("department already exists")
+                # check to see if the course is already listed in the
+                # department
+                dept = dept_list[0]
+                c_list = dept.course.search(number=number)
 
-                    # class already exists in department so don't make a new one
-                    if len(c_list) > 0:
-                        print("class already exists")
-                        c = c_list[0]
+                # class already exists in department so don't make a new one
+                if len(c_list) > 0:
+                    print("class already exists")
+                    c = c_list[0]
 
-                    # the class isn't in the department so make a new one
-                    else:
-                        print("class doens't exist")
-                        c = prereq.Course(number=number).save()
-                        c.dept.connect(dept)
-
-                # the department doesn't exist in the db
+                # the class isn't in the department so make a new one
                 else:
-                    print("department doens't exist")
-                    dept = department.Department(name=dept_str).save()
+                    print("class doens't exist")
                     c = prereq.Course(number=number).save()
-                    dept.uni.connect(uni)
                     c.dept.connect(dept)
 
-                self.prereq.connect(c)
+            # the department doesn't exist in the db
+            else:
+                print("department doens't exist")
+                dept = department.Department(name=dept_str).save()
+                c = prereq.Course(number=number).save()
+                dept.uni.connect(uni)
+                c.dept.connect(dept)
+
+            parent.prereq.connect(c)
 
 
 # using a string here unlike parse_department
@@ -108,37 +109,40 @@ def parse_course(c, m_obj):
     c.credit_hours = int(m_obj.group(4))
 
     prereq_str = m_obj.group(5)
-    added_groups = prereq_str.split('and')
+    print(prereq_str)
 
-    for and_group in added_groups:
-        and_group = and_group.strip()
-        print(and_group)
-        if 'or' in and_group:
-            o = prereq.OrGroup(name=OR_GROUP).save()
+    if prereq_str is not None and prereq_str[:3] != 'Not':
+        added_groups = prereq_str.split('and')
 
-            o.parent.connect(c)
-            add_prereqs(o, and_group)
-            # o.refresh()
-            or_prereqs = o.prereq.all()
+        for and_group in added_groups:
+            and_group = and_group.strip()
+            print(and_group)
+            if 'or' in and_group:
+                o = prereq.OrGroup(name=OR_GROUP).save()
 
-            # if there is more than 1 prereq in the or group
-            # keep the group
-            if len(or_prereqs) > 1:
-                c.prereq.connect(o)
+                o.parent.connect(c)
+                add_prereqs(o, and_group)
+                # o.refresh()
+                or_prereqs = o.prereq.all()
 
-            # if there is only one prereq in the group directly
-            # set that prereq as a prereq to the parent course
-            # and delete the or group
-            elif len(or_prereqs) == 1:
-                add_prereqs(c, and_group)
-                o.delete()
+                # if there is more than 1 prereq in the or group
+                # keep the group
+                if len(or_prereqs) > 1:
+                    c.prereq.connect(o)
 
-            # don't add anything as prereq if there are no
-            # classes in the o
+                # if there is only one prereq in the group directly
+                # set that prereq as a prereq to the parent course
+                # and delete the or group
+                elif len(or_prereqs) == 1:
+                    add_prereqs(c, and_group)
+                    o.delete()
+
+                # don't add anything as prereq if there are no
+                # classes in the o
+                else:
+                    o.delete()
             else:
-                o.delete()
-        else:
-            add_prereqs(c, and_group)
+                add_prereqs(c, and_group)
 
     return c
 
@@ -150,9 +154,10 @@ def parse_course_bulletin(data_str, dept):
     dept.name = ABBREV_DICT[dept_str]
     dept.save()
 
-    it = re.finditer(r"(\d{4})\s+([\w\s,:-]+?)\s+([UG]{1,2})\s+(\d).*?"
-                     "Prereq: (.*?)\.", data_str, re.MULTILINE | re.DOTALL)
+    it = re.finditer(r"(\d{4})\s+([\w\s,:-]+?)\s+([UG]{1,2})\s+(\d).*?</div>.*?</span>"
+                     r"(?:[^<]*?(?:Prereq: (.*?)\.)|[^<]*?)", data_str, re.MULTILINE | re.DOTALL)
 
+    pprint.pprint(it)
     count = 0
     for m_obj in it:
         # here we're passing a match object to parse course
@@ -163,8 +168,8 @@ def parse_course_bulletin(data_str, dept):
 
         count += 1
         # only run 10 times
-        if count > 30:
-            break
+        # if count > 30:
+        # break
 
     return dept
 
@@ -187,9 +192,9 @@ def get_course_bulletin():
 
 
 if __name__ == '__main__':
-    # course_bulletin = dept.get_course_bulletin()
-    f = open('data')
-    course_bulletin = f.read()
+    course_bulletin = get_course_bulletin()
+    # f = open('data')
+    # course_bulletin = f.read()
     parse(course_bulletin)
 
     # f = open('course_bulletin.json', 'w')
